adding copy and vote feature
