+++ app/get-vote/page.tsx
@@ -1,33 +1,40 @@
 "use client";
 
 import { Card, CardContent, Input, Button, Alert } from "@/components/ui";
-import { ExternalLink, Search, Copy, Check } from "lucide-react";
+import { VoteResults, WalletButton, CopyVoteButton, Vote } from "@/components/vote";
+import { useVoteCopy } from "@/hooks/useVoteCopy";
+import { ExternalLink, Search } from "lucide-react";
+import Link from "next/link";
 import { useSearchParams } from "next/navigation";
 import { useEffect, useState, Suspense } from "react";
-
-interface Vote {
-  name: string;
-  address: string;
-  expiry: string;
-  weight: string;
-}
 
 function GetVoteClientContent() {
   const searchParams = useSearchParams();
   const [txHash, setTxHash] = useState("");
   const [loading, setLoading] = useState(false);
   const [votes, setVotes] = useState<Vote[]>([]);
-  const [error, setError] = useState("");
-  const [copied, setCopied] = useState(false);
+  const [fetchError, setFetchError] = useState("");
+
+  const {
+    connectedAddress,
+    sentTxHash,
+    isSendingTx,
+    error: walletError,
+    connectWallet,
+    copyAndSendVote,
+    clearError,
+    signer,
+  } = useVoteCopy();
 
   const handleSubmit = async (hash: string = txHash) => {
     if (!hash) {
-      setError("Please enter a transaction hash");
+      setFetchError("Please enter a transaction hash");
       return;
     }
 
     setLoading(true);
-    setError("");
+    setFetchError("");
+    clearError();
     try {
       const response = await fetch(`/api/get-vote-transaction?txHash=${hash}`);
       if (!response.ok) {
@@ -36,7 +43,7 @@ function GetVoteClientContent() {
       const data = await response.json();
       setVotes(data);
     } catch (error) {
-      setError(
+      setFetchError(
         error instanceof Error ? error.message : "Failed to fetch vote info"
       );
     } finally {
@@ -54,29 +61,30 @@ function GetVoteClientContent() {
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [searchParams]);
 
-  const handleCopyHash = async () => {
-    await navigator.clipboard.writeText(txHash);
-    setCopied(true);
-    setTimeout(() => setCopied(false), 2000);
+  const handleCopyVote = () => {
+    copyAndSendVote(txHash);
   };
 
-  const totalWeight = votes.reduce(
-    (sum, vote) => sum + Number(vote.weight) / 1e18,
-    0
-  );
+  const error = fetchError || walletError;
 
   return (
     <div className="min-h-screen bg-neutral-950">
       {/* Header */}
       <header className="border-b border-neutral-800">
-        <div className="max-w-5xl mx-auto px-6 py-4">
+        <div className="max-w-5xl mx-auto px-6 py-4 flex items-center justify-between">
           <nav className="flex items-center gap-2 text-sm text-neutral-500">
-            <a href="/" className="hover:text-neutral-300 transition-colors">
+            <Link href="/" className="hover:text-neutral-300 transition-colors">
               Home
-            </a>
+            </Link>
             <span>/</span>
             <span className="text-neutral-300">Get Vote</span>
           </nav>
+
+          {/* Wallet Connection */}
+          <WalletButton
+            connectedAddress={connectedAddress}
+            onConnect={connectWallet}
+          />
         </div>
       </header>
 
@@ -88,32 +96,45 @@ function GetVoteClientContent() {
           </h1>
           <p className="text-neutral-400 max-w-lg mx-auto">
             Enter a Pendle vote transaction hash to view the detailed voting
-            information and weight distribution.
+            information and optionally copy it to your own wallet.
           </p>
         </div>
 
         {/* Search Card */}
         <Card className="mb-8">
           <CardContent>
-            <div className="flex flex-col sm:flex-row gap-3">
-              <div className="flex-1">
-                <Input
-                  placeholder="0x..."
-                  value={txHash}
-                  onChange={(e) => setTxHash(e.target.value)}
-                  onKeyDown={(e) => e.key === "Enter" && handleSubmit()}
-                />
+            <div className="flex flex-col gap-4">
+              <div className="flex flex-col sm:flex-row gap-3">
+                <div className="flex-1">
+                  <Input
+                    placeholder="0x..."
+                    value={txHash}
+                    onChange={(e) => setTxHash(e.target.value)}
+                    onKeyDown={(e) => e.key === "Enter" && handleSubmit()}
+                  />
+                </div>
+                <Button
+                  onClick={() => handleSubmit()}
+                  disabled={loading || !txHash}
+                  loading={loading}
+                  loadingText="Fetching..."
+                  className="sm:w-auto w-full"
+                >
+                  <Search className="w-4 h-4 mr-2" />
+                  Lookup
+                </Button>
               </div>
-              <Button
-                onClick={() => handleSubmit()}
-                disabled={loading || !txHash}
-                loading={loading}
-                loadingText="Fetching..."
-                className="sm:w-auto w-full"
-              >
-                <Search className="w-4 h-4 mr-2" />
-                Lookup
-              </Button>
+
+              {votes.length > 0 && (
+                <div className="pt-3 border-t border-neutral-800">
+                  <CopyVoteButton
+                    onCopy={handleCopyVote}
+                    disabled={isSendingTx || !txHash || !signer}
+                    loading={isSendingTx}
+                    hasWallet={!!signer}
+                  />
+                </div>
+              )}
             </div>
           </CardContent>
         </Card>
@@ -125,6 +146,24 @@ function GetVoteClientContent() {
           </Alert>
         )}
 
+        {/* Success Alert */}
+        {sentTxHash && (
+          <Alert variant="success" className="mb-8">
+            <div className="flex flex-col sm:flex-row sm:items-center gap-2">
+              <span>Transaction sent successfully!</span>
+              <a
+                href={`https://etherscan.io/tx/${sentTxHash}`}
+                target="_blank"
+                rel="noopener noreferrer"
+                className="font-mono text-green-400 hover:text-green-300 underline inline-flex items-center gap-1"
+              >
+                {sentTxHash.slice(0, 10)}...
+                <ExternalLink className="w-3 h-3" />
+              </a>
+            </div>
+          </Alert>
+        )}
+
         {/* Loading State */}
         {loading && (
           <div className="text-center py-16">
@@ -133,128 +172,23 @@ function GetVoteClientContent() {
           </div>
         )}
 
-        {/* Results */}
-        {!loading && votes.length > 0 && (
-          <div className="space-y-6">
-            {/* Transaction Info */}
-            <Card>
-              <CardContent className="py-4">
-                <div className="flex flex-col sm:flex-row sm:items-center justify-between gap-4">
-                  <div className="flex items-center gap-3 min-w-0">
-                    <span className="text-sm text-neutral-500 flex-shrink-0">
-                      Transaction:
-                    </span>
-                    <code className="text-sm text-neutral-300 font-mono truncate">
-                      {txHash}
-                    </code>
-                  </div>
-                  <div className="flex items-center gap-2 flex-shrink-0">
-                    <button
-                      onClick={handleCopyHash}
-                      className="p-2 rounded-md hover:bg-neutral-800 text-neutral-400 hover:text-neutral-200 transition-colors"
-                      title="Copy hash"
-                    >
-                      {copied ? (
-                        <Check className="w-4 h-4 text-green-400" />
-                      ) : (
-                        <Copy className="w-4 h-4" />
-                      )}
-                    </button>
-                    <a
-                      href={`https://etherscan.io/tx/${txHash}`}
-                      target="_blank"
-                      rel="noopener noreferrer"
-                      className="p-2 rounded-md hover:bg-neutral-800 text-neutral-400 hover:text-neutral-200 transition-colors"
-                      title="View on Etherscan"
-                    >
-                      <ExternalLink className="w-4 h-4" />
-                    </a>
-                  </div>
-                </div>
-              </CardContent>
-            </Card>
-
-            {/* Stats */}
-            <div className="grid grid-cols-2 gap-4">
-              <Card>
-                <CardContent className="text-center py-6">
-                  <p className="text-3xl font-semibold text-neutral-100">
-                    {votes.length}
-                  </p>
-                  <p className="text-sm text-neutral-500 mt-1">Total Pools</p>
-                </CardContent>
-              </Card>
-              <Card>
-                <CardContent className="text-center py-6">
-                  <p className="text-3xl font-semibold text-neutral-100">
-                    {(totalWeight * 100).toFixed(2)}%
-                  </p>
-                  <p className="text-sm text-neutral-500 mt-1">Total Weight</p>
-                </CardContent>
-              </Card>
-            </div>
-
-            {/* Vote Table */}
-            <Card className="overflow-hidden">
-              <div className="overflow-x-auto">
-                <table className="w-full">
-                  <thead>
-                    <tr className="border-b border-neutral-800 bg-neutral-900/50">
-                      <th className="px-6 py-4 text-left text-xs font-medium text-neutral-500 uppercase tracking-wider">
-                        Pool
-                      </th>
-                      <th className="px-6 py-4 text-left text-xs font-medium text-neutral-500 uppercase tracking-wider">
-                        Address
-                      </th>
-                      <th className="px-6 py-4 text-left text-xs font-medium text-neutral-500 uppercase tracking-wider">
-                        Expiry
-                      </th>
-                      <th className="px-6 py-4 text-right text-xs font-medium text-neutral-500 uppercase tracking-wider">
-                        Weight
-                      </th>
-                    </tr>
-                  </thead>
-                  <tbody className="divide-y divide-neutral-800">
-                    {votes.map((vote, index) => (
-                      <tr
-                        key={index}
-                        className="hover:bg-neutral-800/50 transition-colors"
-                      >
-                        <td className="px-6 py-4">
-                          <span className="text-sm font-medium text-neutral-200">
-                            {vote.name}
-                          </span>
-                        </td>
-                        <td className="px-6 py-4">
-                          <code className="text-xs text-neutral-400 font-mono">
-                            {vote.address.slice(0, 6)}...{vote.address.slice(-4)}
-                          </code>
-                        </td>
-                        <td className="px-6 py-4">
-                          <span className="text-sm text-neutral-400">
-                            {new Date(vote.expiry).toLocaleDateString("en-US", {
-                              year: "numeric",
-                              month: "short",
-                              day: "numeric",
-                            })}
-                          </span>
-                        </td>
-                        <td className="px-6 py-4 text-right">
-                          <span className="text-sm font-medium text-neutral-200">
-                            {((Number(vote.weight) / 1e18) * 100).toFixed(2)}%
-                          </span>
-                        </td>
-                      </tr>
-                    ))}
-                  </tbody>
-                </table>
-              </div>
-            </Card>
+        {/* Sending Transaction State */}
+        {isSendingTx && (
+          <div className="text-center py-16">
+            <div className="inline-block w-10 h-10 border-2 border-neutral-700 border-t-neutral-300 rounded-full animate-spin mb-4" />
+            <p className="text-neutral-500">
+              Sending transaction, please confirm in your wallet...
+            </p>
           </div>
         )}
 
+        {/* Results */}
+        {!loading && !isSendingTx && votes.length > 0 && (
+          <VoteResults votes={votes} txHash={txHash} />
+        )}
+
         {/* Empty State */}
-        {!loading && !error && votes.length === 0 && txHash === "" && (
+        {!loading && !isSendingTx && !error && votes.length === 0 && txHash === "" && (
           <div className="text-center py-16">
             <div className="inline-flex items-center justify-center w-16 h-16 rounded-full bg-neutral-900 mb-4">
               <Search className="w-8 h-8 text-neutral-600" />
