+++ components/vote/index.ts
@@ -1,0 +1,4 @@
+export { VoteResults } from "./VoteResults";
+export type { Vote } from "./VoteResults";
+export { WalletButton } from "./WalletButton";
+export { CopyVoteButton } from "./CopyVoteButton";
